+++ deactivation.php
@@ -10,6 +10,9 @@ declare( strict_types = 1 );
 
 namespace WPGraphQL\Login;
 
+use WPGraphQL\Login\Admin\SettingsRegistry;
+use WPGraphQL\Login\Admin\Upgrade\AbstractUpgrade;
+
 	/**
 	 * Runs when WPGraphQL is de-activated.
 	 *
@@ -42,30 +45,18 @@ function delete_data(): void {
 		return;
 	}
 
-	// Delete plugin version.
-	delete_option( 'wp_graphql_login_version' );
-
 	// Initialize the settings API.
-	$settings = \WPGraphQL\Login\Admin\Settings::get_all_settings();
+	$options = array_merge(
+		[ AbstractUpgrade::VERSION_OPTION_KEY ],
+		array_keys( SettingsRegistry::get_all() ),
+	);
 
-	if ( empty( $settings ) ) {
+	if ( empty( $options ) ) {
 		return;
 	}
 
-	// Get all the setting keys across various groups.
-	$settings = array_reduce(
-		$settings,
-		static function ( $carry, $item ) {
-			return array_merge( $carry, array_keys( $item ) );
-		},
-		[]
-	);
-
-	// Loop over the registered settings fields and delete the options.
-	if ( ! empty( $settings ) && is_array( $settings ) ) {
-		foreach ( array_keys( $settings ) as $option_name ) {
-			delete_option( $option_name );
-		}
+	foreach ( $options as $option ) {
+		delete_option( $option );
 	}
 
 	// Fire an action when data is deleted.