+++ global.d.ts
@@ -3,12 +3,17 @@
 declare let __webpack_public_path__: string;
 
 declare module '*.svg' {
-  import * as React from 'react';
+	import * as React from 'react';
 
-  export const ReactComponent: React.FunctionComponent<
-    React.SVGProps<SVGSVGElement> & { title?: string }
-  >;
+	export const ReactComponent: React.FunctionComponent<
+		React.SVGProps< SVGSVGElement > & { title?: string }
+	>;
 
-  const src: string;
-  export default src;
+	const src: string;
+	export default src;
+}
+
+declare module '*.scss' {
+	const content: Record< string, string >;
+	export default content;
 }