+++ packages/admin/components/ui/loading/index.tsx
@@ -1,0 +1,11 @@
+import { Spinner } from '@wordpress/components';
+import clsx from 'clsx';
+
+export const Loading = ( {
+	className,
+	...rest
+}: React.ComponentProps< typeof Spinner > ) => {
+	const classes = clsx( 'wp-graphql-headless-login__loading', className );
+
+	return <Spinner className={ classes } { ...rest } />;
+};