+++ packages/admin/settings/index.ts
@@ -1,3 +1,0 @@
-export * from './AccessControlSettings';
-export * from './ClientSettings';
-export * from './PluginSetings';