+++ packages/admin/types.d.ts
@@ -3,16 +3,10 @@ import { _Hooks } from '@wordpress/hooks/build-types/createHooks';
 declare global {
 	const wpGraphQLLogin: {
 		hooks: _Hooks;
-		settings: {
-			accessControl: Record<
-				keyof AccessControlSettingsType,
-				SettingSchema
-			>;
-			providers: Record<
-				string,
-				Record<keyof ProviderSettingType, SettingSchema>
-			>;
-			plugin: Record<keyof PluginSettingsType, SettingSchema>;
+		settings: SettingSchema & {
+			providers: Record< string, Record< string, FieldSchema & {
+				properties: Record< string, FieldSchema >;
+			} > >;
 		};
 		nonce: string;
 		secret: {
@@ -22,41 +16,46 @@ declare global {
 	};
 }
 
-type SettingSchema = {
-	advanced?: boolean;
+type AllowedSettingKeys =
+	| 'providers'
+	| 'wpgraphql_login_settings'
+	| 'wpgraphql_login_access_control';
+
+type AllowedConditionalLogicOperators = '==' | '!=' | '>' | '<' | '>=' | '<=';
+
+type FieldSchema = {
+	description: string;
+	label: string;
+	type: string;
+	conditionalLogic?: [
+		{
+			slug: string;
+			operator: AllowedConditionalLogicOperators;
+			value: string | number | boolean;
+		},
+	];
+	controlOverrides?: Record< string, unknown >;
+	controlType?: string;
 	default?: unknown;
-	description?: string;
+	enum?: string[]; // @todo Not sure if this is the correct type
 	help?: string;
 	hidden?: boolean;
-	label?: string;
+	isAdvanced?: boolean;
 	order?: number;
 	required?: boolean;
-	type?: string;
-	[key: string]: any;
 };
 
-type AccessControlSettingsType = {
-	hasAccessControlAllowCredentials?: boolean;
-	hasSiteAddressInOrigin?: boolean;
-	additionalAuthorizedDomains?: string[];
-	shouldBlockUnauthorizedDomains?: boolean;
-	customHeaders?: string[];
-	[key: string]: any;
-};
-
-type PluginSettingsType = {
-	wpgraphql_login_settings_show_avanced_settings?: boolean;
-	wpgraphql_login_settings_jwt_secret_key?: string;
-	wpgraphql_login_settings_delete_data_on_deactivate?: boolean;
-	[key: string]: any;
-};
+type SettingSchema = Record<
+	AllowedSettingKeys,
+	Record< string, FieldSchema >
+>;
 
 type ProviderSettingType = {
 	name: string;
 	order: number;
 	slug?: string;
 	isEnabled: boolean;
-	[key: string]: any;
+	[ key: string ]: any;
 	clientOptions: ClientOptionsType;
 	loginOptions: LoginOptionsType;
 };
@@ -69,24 +68,27 @@ type OAuth2ClientOptionsType = Record<
 		redirectUri: string;
 		clientId: string;
 		clientSecret: string;
-		[key: string]: any;
+		[ key: string ]: any;
 	}
 >;
 
-type SiteTokenClientOptionsType = Record<string,{
-	headerKey: string;
-	secretKey: string;
-	[key: string]: any;
-}>
+type SiteTokenClientOptionsType = Record<
+	string,
+	{
+		headerKey: string;
+		secretKey: string;
+		[ key: string ]: any;
+	}
+>;
 
 type LoginOptionsType = {
 	useAuthenticationCookie?: boolean;
-	[key: string]: any;
-} & (OAuth2LoginOptionsType | SiteTokenLoginOptionsType);
+	[ key: string ]: any;
+} & ( OAuth2LoginOptionsType | SiteTokenLoginOptionsType );
 
 type SiteTokenLoginOptionsType = {
 	metaKey?: string;
-	[key: string]: any;
+	[ key: string ]: any;
 };
 
 type OAuth2LoginOptionsType = {