+++ tests/_data/config.php
@@ -1,6 +1,1 @@
 <?php //phpcs:ignoreFile
-
-// Fixes compatibility with WP6.0
-if ( ! isset( $GLOBALS['wp_filters'] ) ) {
-	$GLOBALS['wp_filters'] = [];
-}