+++ tests/wpunit/MainTest.php
@@ -24,7 +24,7 @@ class MainTest extends \lucatume\WPBrowser\TestCase\WPTestCase {
 	 */
 	public function testInstance() {
 		$this->instance = new Main();
-		$this->assertTrue( $this->instance instanceof Main );
+		$this->assertInstanceOf( Main::class, $this->instance );
 	}
 
 	/**
@@ -38,14 +38,27 @@ class MainTest extends \lucatume\WPBrowser\TestCase\WPTestCase {
 	}
 
 	/**
-	 * @covers \WPGraphQL\Login\Main::__wakeup
-	 * @covers \WPGraphQL\Login\Main::__clone
+	 * Test cloning does not work.
+	 *
+	 * @covers \WPGraphQL\Login\Main
 	 */
-	public function testClone() {
-		$actual = Main::instance();
-		$rc     = new ReflectionClass( $actual );
-		$this->assertTrue( $rc->hasMethod( '__clone' ) );
-		$this->assertTrue( $rc->hasMethod( '__wakeup' ) );
+	public function testClone(): void {
+		$this->setExpectedIncorrectUsage( '__clone' );
+
+		$instance = Main::instance();
+		clone $instance;
+	}
+
+	/**
+	 * Test deserializing does not work.
+	 */
+	public function testWakeup(): void {
+		$this->setExpectedIncorrectUsage( '__wakeup' );
+
+		$instance            = Main::instance();
+		$serialized_instance = serialize( $instance );
+
+		unserialize( $serialized_instance );
 	}
 
 	/**